+++ src/components/ui/PostList.tsx
@@ -1,12 +1,11 @@
 import { Flex, Stack, StackProps, Text } from '@mantine/core';
 import { IPost } from '../../types/models';
 import { ApiError, SuccessWithPagination } from '../../types/api';
-import WithStatus from '../layout/WithStatus';
 import Post from './Post';
-import WithDataHandler from '../layout/WithDataHandler';
 import { InfiniteData } from '@tanstack/react-query';
 import Infinite from '../layout/Infinite';
 import SortButton from '../buttons/SortButton';
+import { useData } from '../../hooks/useData';
 
 interface PostListProps extends StackProps {
   data: InfiniteData<SuccessWithPagination<IPost>> | undefined;
@@ -17,37 +16,33 @@ interface PostListProps extends StackProps {
   place: 'feed' | 'profile';
 }
 
-const PostList = (props: PostListProps) => {
-  const { data, hasNext, fetchNext, ...rest } = props;
-
-  const total = data?.pages[0].meta.total_items ?? 0;
+const PostList = ({ data, status, error, hasNext, fetchNext, place, ...props }: PostListProps) => {
+  const { Element, total } = useData({
+    data,
+    status,
+    noContentLabel: 'No posts',
+  });
 
   return (
-    <Stack gap={0} {...rest}>
-      {total === 0 ? (
-        <Text ta='center' c='dimmed'>
-          No Posts
-        </Text>
-      ) : (
-        <>
-          <Flex justify='space-between' align='center'>
-            {data && (
-              <Text fz='xs' c='dimmed'>
-                {total} Total {total === 1 ? 'Post' : 'Posts'}
-              </Text>
-            )}
-            {total !== 0 && <SortButton place={props.place} />}
-          </Flex>
-          <Infinite
-            data={data}
-            hasNext={hasNext}
-            fetchNext={fetchNext}
-            Item={({ item }) => <Post post={item} />}
-          />
-        </>
-      )}
-    </Stack>
+    Element ?? (
+      <Stack gap={0} {...props}>
+        <Flex justify='space-between' align='center'>
+          {data && (
+            <Text fz='xs' c='dimmed'>
+              {total} Total {total === 1 ? 'Post' : 'Posts'}
+            </Text>
+          )}
+          {total !== 0 && <SortButton place={place} />}
+        </Flex>
+        <Infinite
+          data={data}
+          hasNext={hasNext}
+          fetchNext={fetchNext}
+          Item={({ item }) => <Post post={item} />}
+        />
+      </Stack>
+    )
   );
 };
 
-export default WithStatus(WithDataHandler(PostList, 'No posts found'));
+export default PostList;