+++ src/components/ui/ProfileDesc.tsx
@@ -1,6 +1,6 @@
 import { Text, useMantineTheme } from '@mantine/core';
 import { Flex, Group } from '@mantine/core';
-import { IconCake } from '@tabler/icons-react';
+import { IconCake, IconUserPlus } from '@tabler/icons-react';
 import dayjs from 'dayjs';
 import { IUser } from '../../types/models';
 
@@ -10,10 +10,16 @@ const ProfileDesc = ({ user }: { user: IUser }) => {
   } = useMantineTheme();
 
   return (
-    <Flex justify='flex-end'>
+    <Flex justify='space-between'>
       <Group align='center' gap={10}>
-        <IconCake size={15} color={gray[7]} />
-        <Text fz='sm' c='dimmed'>
+        <IconCake size={13} color={gray[7]} />
+        <Text fz='xs' c='dimmed'>
+          Born on {dayjs(user.date_of_birth).format('D MMMM YYYY')}
+        </Text>
+      </Group>
+      <Group align='center' gap={10}>
+        <IconUserPlus size={13} color={gray[7]} />
+        <Text fz='xs' c='dimmed'>
           Joined on {dayjs(user.created_at).format('D MMMM YYYY')}
         </Text>
       </Group>